+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Click up arrow to go back to TOC.
 - Change the replacement condition of the audio_split_cmd only when the user doesn't modify it.
 - Change the MAX_REGION_SIZE_LIMIT into 60 seconds.
 - Change all text file input decoding into "utf-8".
+- Change the default style selection in subtitles translation.
 
 #### Fixed(Unreleased)